refactor: 더미 스트림 생성 함수 외부로 분리
